backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
langchain>=0.1.6
langchain-openai>=0.0.5
langchain-community>=0.0.19
azure-search-documents==11.4.0
azure-core>=1.29.5
beautifulsoup4>=4.12.3
requests>=2.31.0
tiktoken>=0.5.2
python-dotenv>=1.0.1
gunicorn>=21.2.0
azure-cognitiveservices-speech>=1.38.0
python-telegram-bot>=20.7
